refactor: finish moving rest of utils into subfolder
