+++ src/rules/utils/accessors.ts
@@ -1,20 +1,4 @@
-import { parse as parsePath } from 'path';
-import {
-  AST_NODE_TYPES,
-  ESLintUtils,
-  TSESLint,
-  TSESTree,
-} from '@typescript-eslint/utils';
-import { version } from '../../package.json';
-import { isTypeOfJestFnCall } from './utils/parseJestFnCall';
-
-const REPO_URL = 'https://github.com/jest-community/eslint-plugin-jest';
-
-export const createRule = ESLintUtils.RuleCreator(name => {
-  const ruleName = parsePath(name).name;
-
-  return `${REPO_URL}/blob/v${version}/docs/rules/${ruleName}.md`;
-});
+import { AST_NODE_TYPES, TSESTree } from '@typescript-eslint/utils';
 
 /**
  * A `Literal` with a `value` of type `string`.
@@ -109,54 +93,6 @@ export const getStringValue = <S extends string>(node: StringNode<S>): S =>
   isTemplateLiteral(node) ? node.quasis[0].value.raw : node.value;
 
 /**
- * Represents a `MemberExpression` with a "known" `property`.
- */
-export interface KnownMemberExpression<Name extends string = string>
-  extends TSESTree.MemberExpressionComputedName {
-  property: AccessorNode<Name>;
-}
-
-/**
- * Represents a `CallExpression` with a "known" `property` accessor.
- *
- * i.e `KnownCallExpression<'includes'>` represents `.includes()`.
- */
-export interface KnownCallExpression<Name extends string = string>
-  extends TSESTree.CallExpression {
-  callee: CalledKnownMemberExpression<Name>;
-}
-
-/**
- * Represents a `MemberExpression` with a "known" `property`, that is called.
- *
- * This is `KnownCallExpression` from the perspective of the `MemberExpression` node.
- */
-interface CalledKnownMemberExpression<Name extends string = string>
-  extends KnownMemberExpression<Name> {
-  parent: KnownCallExpression<Name>;
-}
-
-/**
- * Represents a `CallExpression` with a single argument.
- */
-export interface CallExpressionWithSingleArgument<
-  Argument extends TSESTree.Expression = TSESTree.Expression,
-> extends TSESTree.CallExpression {
-  arguments: [Argument];
-}
-
-/**
- * Guards that the given `call` has only one `argument`.
- *
- * @param {CallExpression} call
- *
- * @return {call is CallExpressionWithSingleArgument}
- */
-export const hasOnlyOneArgument = (
-  call: TSESTree.CallExpression,
-): call is CallExpressionWithSingleArgument => call.arguments.length === 1;
-
-/**
  * An `Identifier` with a known `name` value - i.e `expect`.
  */
 interface KnownIdentifier<Name extends string> extends TSESTree.Identifier {
@@ -231,77 +167,3 @@ export const getAccessorValue = <S extends string = string>(
 export type AccessorNode<Specifics extends string = string> =
   | StringNode<Specifics>
   | KnownIdentifier<Specifics>;
-
-export enum DescribeAlias {
-  'describe' = 'describe',
-  'fdescribe' = 'fdescribe',
-  'xdescribe' = 'xdescribe',
-}
-
-export enum TestCaseName {
-  'fit' = 'fit',
-  'it' = 'it',
-  'test' = 'test',
-  'xit' = 'xit',
-  'xtest' = 'xtest',
-}
-
-export enum HookName {
-  'beforeAll' = 'beforeAll',
-  'beforeEach' = 'beforeEach',
-  'afterAll' = 'afterAll',
-  'afterEach' = 'afterEach',
-}
-
-const joinNames = (a: string | null, b: string | null): string | null =>
-  a && b ? `${a}.${b}` : null;
-
-export function getNodeName(node: TSESTree.Node): string | null {
-  if (isSupportedAccessor(node)) {
-    return getAccessorValue(node);
-  }
-
-  switch (node.type) {
-    case AST_NODE_TYPES.TaggedTemplateExpression:
-      return getNodeName(node.tag);
-    case AST_NODE_TYPES.MemberExpression:
-      return joinNames(getNodeName(node.object), getNodeName(node.property));
-    case AST_NODE_TYPES.NewExpression:
-    case AST_NODE_TYPES.CallExpression:
-      return getNodeName(node.callee);
-  }
-
-  return null;
-}
-
-export type FunctionExpression =
-  | TSESTree.ArrowFunctionExpression
-  | TSESTree.FunctionExpression;
-
-export const isFunction = (node: TSESTree.Node): node is FunctionExpression =>
-  node.type === AST_NODE_TYPES.FunctionExpression ||
-  node.type === AST_NODE_TYPES.ArrowFunctionExpression;
-
-export const getTestCallExpressionsFromDeclaredVariables = (
-  declaredVariables: readonly TSESLint.Scope.Variable[],
-  scope: TSESLint.Scope.Scope,
-): TSESTree.CallExpression[] => {
-  return declaredVariables.reduce<TSESTree.CallExpression[]>(
-    (acc, { references }) =>
-      acc.concat(
-        references
-          .map(({ identifier }) => identifier.parent)
-          .filter(
-            (node): node is TSESTree.CallExpression =>
-              node?.type === AST_NODE_TYPES.CallExpression &&
-              isTypeOfJestFnCall(node, scope, ['test']),
-          ),
-      ),
-    [],
-  );
-};
-
-export * from './utils/parseJestFnCall';
-export * from './utils/detectJestVersion';
-export * from './utils/parseExpectCall';
-export * from './utils/followTypeAssertionChain';
